ml/requirements.txt
fastapi
uvicorn
pillow
numpy
opencv-python
scikit-image